Lexing identifiers
Summary:
This isn't the greatest, the is_alpha function is checking for
non-reserved characters. Because jsonnet is utf-8, the identifier can
be any utf-8 valid variable width character. So rather than define the
byte range -- which hmm -- this uses a "negative set" approach. It
references the reserved characters that cannot be part of a valid
utf-8 Jsonnet string and checks against those.