Some rudimentary parsing
What:
Ok so this does a parse of a unary, followed by a literal. Idea is to
provide the proper order of parsing -- unary first then primary. Each
one of these rules can only return a specific type of expression so we
put the expression in the vector that matches the type and return the
index of that. Probably should return a tuple of index and type.