Recurse into lower rules
Summary:
As this expands to more and more rules, the higher priority rules need
to be abe to call the lower rules and return the tuple. In particular,
in the main parsing routine, it needs to call the highest priority
rule and have it return the results it finds. Prior to this change, it
called the higher priority rule and if it returned None, then called
the next one in order. This probably wouldn't have worked with a
reasonably complex parsing input

Now the higher order rules call into the lower order rules when it
doesn't find a match.