Implement Shift parsing
Summary:
The shift operators are lower in parse than comparsion operators and
are next on the list to write.

Operators <<, >>, &, ^, | and ~ first convert their operands to signed
64 bit integers, then perform the operations in a standard way, then
convert back to IEEE double precision floating point. In shift
operations <<, >>, the right hand value modulo 64 is interpreted as
the shift count. Shifting with a negative shift count raises an error.

https://jsonnet.org/ref/spec.html