Start of Error handling in parser
Summary:
Need to collect errors at different points. And when the errors are
present, we need to error out when rendering the tree

And need to keep parsing until the end -- and in the case of jsonnet,
semicolon at the end is invalid. This adds some logic to keep parsing
until EOF and asserts that EOF is hit for an valid AST.